+++ CHANGELOG.md
@@ -14,6 +14,7 @@ keys in the response body.
 
 - server: fix issue of not exposing mutation functions to the admin role when function permissions are inferred (fix #6503)
 - server: add "resource_version" field to metadata for concurrency control - disable lookup during migrations
+- server: fix issue with queries on character column types (close #6217)
 - server: optimize resolving source. Resolving a source would create connection pools every time. Optimize that to re-create connection pools only when necessary. (#609)
 
 ## v1.4.0-alpha.1